requirements.txt
braintree==3.59.0
celery==4.4.2
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
channels-redis==2.4.2
cycler==0.10.0
daphne==2.5.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1
django-braces==1.14.0
django-embed-video==1.3.2
django-extensions==2.2.5
django-memcache-status==2.2
django-parler==2.0.1
django-phonenumber-field==5.0.0
django-rosetta==0.9.3
django-taggit==1.2.0
djangorestframework==3.11.0
duplicity==0.8.12.0
easy-thumbnails==2.7
entrypoints==0.3
fasteners==0.14.1
flower==0.9.3
gunicorn==20.0.4
incremental==17.5.0
ipykernel==5.3.4
ipywidgets==7.5.1
isort==5.4.2
Markdown==3.2.1
MarkupSafe==1.1.0
phonenumbers==8.12.13
pickleshare==0.7.5
Pillow==7.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.4
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==3.141.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.25.8
uWSGI==2.0.18
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
